Create explicit dependency on the postpone filter.
This fixes an issues where subrequests wouldn't have their order preserved and
the output chain would be mangled. It was assumed this related to the SSI module
which has dependency on the postpone filter and was including it in affected cases.

By making this dependency explicit, we no longer require the SSI module for this
to work.